# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: SoccerCoachKit API
  description: >-
    Backend for SoccerCoachKit — a coach → club evaluation platform. Models the
    whole castle (organizations, roles, persons, guardianships, time-bounded
    roster memberships, a generic evaluation engine, polymorphic sharing) while
    Phase 1 ships the solo-coach surface: identity, teams + roster, and the
    pre/post-game evaluation loop with cross-instance score aggregation.
  version: 0.2.0
servers:
  - url: /
    description: Current host
...
paths:
...
  /api/v1/teams:
    get:
      tags: [teams]
      summary: List teams in your organization
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/OrgHeader" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Team" } }
    post:
      tags: [teams]
      summary: Create a team
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/OrgHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTeamRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
  /api/v1/teams/{id}:
    get:
      tags: [teams]
      summary: Get a team with its active roster
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  team: { $ref: "#/components/schemas/Team" }
                  roster: { type: array, items: { $ref: "#/components/schemas/RosterEntry" } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [teams]
      summary: Delete a team
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    OrgHeader:
      in: header
      name: X-Organization-ID
      required: false
      description: Act within a specific organization; defaults to your single/first org.
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...
    Team:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        name: { type: string }
        ageGroup: { type: string, nullable: true }
        season: { type: string, nullable: true }
        activeRosterCount: { type: integer }
        createdAt: { type: string, format: date-time }
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        ageGroup: { type: string }
        season: { type: string }
    RosterEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        personId: { type: string, format: uuid }
        displayName: { type: string }
        email: { type: string, nullable: true }
        birthdate: { type: string, nullable: true }
        jerseyNumber: { type: integer, nullable: true }
        position: { type: string, nullable: true }
        joinedOn: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive] }
...